server,setup: add server.New()
Right now you kind of are on your own for configuring the database and
then the server. Add a New() method with a Config{} that will allow us
to put more configuration on it and also internalize more of the
configuration, so the user doesn't have to do as much.