all: add Handler interface for dequeuer
A Handler allows more flexibility about how jobs are processed - they
don't need to be sent to a downstream server using the existing
client.

Use a context.Context for cancelation in more places, instead of
relying on a fixed Timeout in the HTTP client. The dequeuer no longer
has a QuitChan for example - use a context.Context to cancel stuff
instead.

Update some tests to handle new behavior - the context.Context means
we lose the ability to introspect the phase of the HTTP request we
were at, which is probably fine.