dequeuer: simplify configuration with WorkServer
See the Example or commands/dequeuer for an example of how to use it.
It should be a lot easier to configure this in the common case now.

All of the metrics queries accept a context.Context now, which means
we can cancel them when we cancel the parent WorkServer.

Replace gorilla/handlers.LoggingServer with kevinburke/handlers.Log.